+++ components/calendar/src/hebrew.rs
@@ -274,13 +274,13 @@ impl Calendar for Hebrew {
         if is_leap_year {
             if ordinal == 6 {
                 return types::MonthInfo {
-                    ordinal: ordinal as u32,
+                    ordinal: ordinal,
                     standard_code: types::MonthCode(tinystr!(4, "M05L")),
                     formatting_code: types::MonthCode(tinystr!(4, "M05L")),
                 };
             } else if ordinal == 7 {
                 return types::MonthInfo {
-                    ordinal: ordinal as u32,
+                    ordinal: ordinal,
                     // Adar II is the same as Adar and has the same code
                     standard_code: types::MonthCode(tinystr!(4, "M06")),
                     formatting_code: types::MonthCode(tinystr!(4, "M06L")),
@@ -309,7 +309,7 @@ impl Calendar for Hebrew {
         };
 
         types::MonthInfo {
-            ordinal: date.0.month as u32,
+            ordinal: date.0.month,
             standard_code: types::MonthCode(code),
             formatting_code: types::MonthCode(code),
         }
